fix: fix sagas test
