+++ packages/blockchain-wallet-v4-frontend/src/data/components/debitCard/sagas.test.js
@@ -1,0 +1,146 @@
+import { call, put, select } from 'redux-saga/effects'
+
+import * as S from 'data/components/debitCard/selectors'
+
+import sagas from './sagas'
+import { actions as A } from './slice'
+
+const API_MOCK = {
+  getDCCurrentAccount: jest.fn()
+}
+
+jest.mock('data/modules/profile/sagas', () => () => ({
+  waitForUserData: jest.fn()
+}))
+
+describe('debitCard sagas', () => {
+  const api = API_MOCK
+  const coreSagas = null
+  const networks = null
+  const CARD_ID = 'some-id'
+
+  const debitCardSagas = sagas({ api, coreSagas, networks })
+
+  describe('getCurrentCardAccount', () => {
+    const saga = debitCardSagas.getCurrentCardAccount
+    const currentAccountSelectedMock = { accountCurrency: 'USD' }
+    const accountMock = { balance: { symbol: 'USD', value: '1' } }
+    const eligibleAccountsMock = []
+    eligibleAccountsMock.push(accountMock)
+
+    const validateYieldSelectGetEligibleAccounts = (generator) => {
+      const actual = generator.next().value
+      const expected = select(S.getEligibleAccounts)
+
+      expect(actual).toEqual(expected)
+    }
+    const validateYieldPutGetCurrentCardAccountLoading = (generator) => {
+      const actual = generator.next(eligibleAccountsMock).value
+      const expected = put(A.getCurrentCardAccountLoading())
+
+      expect(actual).toEqual(expected)
+    }
+    const validateYieldCallGetDCCurrentAccount = (generator) => {
+      const actual = generator.next().value
+
+      const expected = call(api.getDCCurrentAccount, CARD_ID)
+      expect(actual).toEqual(expected)
+    }
+
+    describe('success flow with currentAccount data', () => {
+      const generator = saga(CARD_ID)
+
+      it('should yields select getEligibleAccounts', () => {
+        validateYieldSelectGetEligibleAccounts(generator)
+      })
+
+      it('should yields put getCurrentCardAccountLoading', () => {
+        validateYieldPutGetCurrentCardAccountLoading(generator)
+      })
+
+      it('should yield call getDCCurrentAccount', () => {
+        validateYieldCallGetDCCurrentAccount(generator)
+      })
+
+      it('should yield put getCurrentCardAccountSuccess', () => {
+        const actual = generator.next(currentAccountSelectedMock).value
+        const expected = put(A.getCurrentCardAccountSuccess(accountMock))
+
+        expect(actual).toEqual(expected)
+      })
+
+      it('should be done', () => {
+        const actual = generator.next().done
+        const expected = true
+
+        expect(actual).toEqual(expected)
+      })
+    })
+
+    describe('no current account obtained from API', () => {
+      const generator = saga(CARD_ID)
+
+      it('should yields select getEligibleAccounts', () => {
+        validateYieldSelectGetEligibleAccounts(generator)
+      })
+
+      it('should yields put getCurrentCardAccountLoading', () => {
+        validateYieldPutGetCurrentCardAccountLoading(generator)
+      })
+
+      it('should yield call getDCCurrentAccount', () => {
+        validateYieldCallGetDCCurrentAccount(generator)
+      })
+
+      it('should yield put getCurrentCardAccountFailure', () => {
+        const noCurrentAccountResponse = { accountCurrency: null }
+        const actual = generator.next(noCurrentAccountResponse).value
+        const expected = put(
+          A.getCurrentCardAccountFailure('Could not get current user funds, no_funds_obtained')
+        )
+
+        expect(actual).toEqual(expected)
+      })
+
+      it('should be done', () => {
+        const actual = generator.next().done
+        const expected = true
+
+        expect(actual).toEqual(expected)
+      })
+    })
+
+    describe('current account mismatch with existing eligible accounts', () => {
+      const generator = saga(CARD_ID)
+
+      it('should yields select getEligibleAccounts', () => {
+        validateYieldSelectGetEligibleAccounts(generator)
+      })
+
+      it('should yields put getCurrentCardAccountLoading', () => {
+        validateYieldPutGetCurrentCardAccountLoading(generator)
+      })
+
+      it('should yield call getDCCurrentAccount', () => {
+        validateYieldCallGetDCCurrentAccount(generator)
+      })
+
+      it('should yield put getCurrentCardAccountFailure', () => {
+        const noCurrentAccountResponse = { accountCurrency: 'ARS' }
+        const actual = generator.next(noCurrentAccountResponse).value
+        const expected = put(
+          A.getCurrentCardAccountFailure('Could not get current user funds, no_funds_obtained')
+        )
+
+        expect(actual).toEqual(expected)
+      })
+
+      it('should be done', () => {
+        const actual = generator.next().done
+        const expected = true
+
+        expect(actual).toEqual(expected)
+      })
+    })
+  })
+})
