+++ packages/blockchain-wallet-v4-frontend/src/data/components/debitCard/sagas.ts
@@ -47,17 +47,15 @@ export default ({ api, coreSagas, networks }: { api: APIType; coreSagas: any; ne
 
       const { accountCurrency } = yield call(api.getDCCurrentAccount, cardId)
 
-      if (!accountCurrency && isEmpty(eligibleAccounts)) throw new Error('no_funds_obtained')
+      const accountFound = findAccount(accountCurrency, eligibleAccounts)
 
-      yield put(A.getCurrentCardAccountSuccess(findAccount(accountCurrency, eligibleAccounts)))
+      if (!accountCurrency || !accountFound) throw new Error('no_funds_obtained')
+
+      yield put(A.getCurrentCardAccountSuccess(accountFound))
     } catch (e) {
-      console.error('Failed to get current card account', e)
-      if (!isEmpty(eligibleAccounts)) {
-        // In case of failure it is set the default account as current
-        yield put(A.getCurrentCardAccountSuccess(eligibleAccounts[0]))
-      } else {
-        yield put(A.getCurrentCardAccountFailure('Could not get user funds'))
-      }
+      yield put(
+        A.getCurrentCardAccountFailure(`Could not get current user funds, ${errorHandler(e)}`)
+      )
     }
   }
